Move vds writer out of the data writing mess
