+++ src/nexgen/tools/VDS_tools.py
@@ -1,0 +1,58 @@
+"""
+Tools to write Virtual DataSets
+"""
+
+import h5py
+import logging
+
+from pathlib import Path
+from typing import List
+
+vds_logger = logging.getLogger("NeXusGenerator.writer.vds")
+
+
+def vds_writer(nxsfile: h5py.File, datafiles: List[Path], vds_writer: str):
+    """
+    Write a Virtual Dataset file for image data.
+
+    Args:
+        nxsfile:    NeXus file being written.
+        datafiles:  List of paths to source files.
+        vds_writer: Choose whether to write a virtual dataset under /entry/data/data in the NeXus file
+                    or a _vds.h5 file added to the NeXus file as an External Link.
+    """
+    vds_logger.info("Start creating VDS ...")
+    entry_key = "data"
+    # Calculate total number of frames across the files
+    frames = [h5py.File(f, "r")[entry_key].shape[0] for f in datafiles]
+    tot_frames = sum(frames)
+    # Get shape of the detector
+    sh = h5py.File(datafiles[0], "r")[entry_key].shape[1:]
+
+    dtyp = h5py.File(datafiles[0], "r")[entry_key].dtype
+
+    # Create virtual layout
+    layout = h5py.VirtualLayout(shape=(tot_frames,) + sh, dtype=dtyp)
+    start = 0
+    for n, filename in enumerate(datafiles):
+        end = start + frames[n]
+        vsource = h5py.VirtualSource(
+            filename.name, entry_key, shape=(frames[n],) + sh
+        )  # Source definition
+        layout[start:end:1, :, :] = vsource
+        start = end
+
+    if vds_writer == "dataset":
+        # Write virtual dataset in nexus file
+        nxdata = nxsfile["entry/data"]
+        nxdata.create_virtual_dataset(entry_key, layout, fillvalue=-1)
+        vds_logger.info("VDS written to NeXus file.")
+    elif vds_writer == "file":
+        # Create a _vds.h5 file and add link to nexus file
+        s = Path(nxsfile.filename).expanduser().resolve()
+        vds_filename = s.parent / f"{s.stem}_vds.h5"
+        del s
+        with h5py.File(vds_filename, "w") as vds:
+            vds.create_virtual_dataset("data", layout, fillvalue=-1)
+        nxsfile["entry/data/data"] = h5py.ExternalLink(vds_filename.name, "data")
+        vds_logger.info(f"{vds_filename} written and link added to NeXus file.")
